Sony pdx206: Use correct power_profile from stock
I've experienced one completely weird bug with GSIs on my Xperia 5 II (pdx206) depending on battery,  which have shown wrong capacities (something around 3000mAh and 3200mAh).

On hardware side, it's 4000mAh, but Sony seems to limit it to 3860mAh, maybe for battery protection. We should keep it as is.

So this commit implements the right power_profile for pdx206 devices.